LPS-3399
git-svn-id: svn://svn.liferay.com/repos/public/portal/trunk@31716 05bdf26c-840f-0410-9ced-eb539d925f36

+++ portal-impl/src/com/liferay/portal/util/FileImpl.java
@@ -205,7 +205,7 @@ public class FileImpl implements com.liferay.portal.kernel.util.File {
 	}
 
 	public boolean delete(File file) {
-		if (file.exists()) {
+		if ((file != null) && file.exists()) {
 			return file.delete();
 		}
 		else {
